Use redirector before enrollment
Previously the script enrolled the device with CV using
Terminattr before using the redirector. This worked with
newer versions of Terminattr because Terminattr
automatically contacted the redirector internally to
redirect the enrollment to the correct cluster.

However this didn't work on older Terminattr releases
because those releases don't know to contact the redirector.
This change contacts the redirector first so that Terminattr
doesn't need to.